feat: migrate project for typescript structure
