+++ index.js
@@ -3,7 +3,7 @@
  */
 
 import { AppRegistry } from 'react-native'
-import App from 'src/App'
-import { name as appName } from 'test_app/app.json'
+import App from './src'
+import { name as appName } from './app.json'
 
 AppRegistry.registerComponent(appName, () => App)
