+++ tslint.json
@@ -1,0 +1,143 @@
+{
+    "extends": [
+        "tslint:recommended",
+        "tslint-config-standard",
+        "tslint-react",
+        "tslint-config-prettier"
+    ],
+    "rules": {
+        "align": [
+            true,
+            "parameters",
+            "statements"
+        ],
+        "ban": false,
+        "ban-types": false,
+        "class-name": true,
+        "comment-format": [
+            true,
+            "check-space"
+        ],
+        "curly": false,
+        "forin": false,
+        "indent": [
+            true,
+            "spaces"
+        ],
+        "no-object-literal-type-assertion": false,
+        "no-angle-bracket-type-assertion": false,
+        "jsdoc-format": false,
+        "jsx-boolean-value": false,
+        "jsx-curly-spacing": true,
+        "jsx-no-multiline-js": false,
+        "jsx-no-lambda": false,
+        "jsx-no-string-ref": false,
+        "label-position": true,
+        "max-classes-per-file": false,
+        "max-line-length": [
+            false,
+            140
+        ],
+        "member-access": false,
+        "member-ordering": [
+            true,
+            {
+                "order": [
+                    "private-static-field",
+                    "public-static-field",
+                    "protected-static-field",
+                    "public-static-method",
+                    "private-static-method",
+                    "protected-static-method",
+                    "public-instance-field",
+                    "private-instance-field",
+                    "protected-instance-field",
+                    "public-constructor",
+                    "private-constructor",
+                    "protected-constructor"
+                ]
+            }
+        ],
+        "no-any": false,
+        "no-arg": false,
+        "no-bitwise": false,
+        "no-conditional-assignment": true,
+        "no-console": true,
+        "no-submodule-imports": false,
+        "no-construct": true,
+        "no-debugger": true,
+        "no-duplicate-variable": true,
+        "no-empty": true,
+        "no-empty-interface": false,
+        "interface-name": [true],
+        "no-eval": true,
+        "no-implicit-dependencies": false,
+        "no-namespace": false,
+        "no-inferrable-types": false,
+        "no-internal-module": true,
+        "no-require-imports": false,
+        "no-shadowed-variable": false,
+        "no-string-literal": true,
+        "no-switch-case-fall-through": true,
+        "no-trailing-whitespace": true,
+        "no-unused-expression": true,
+        "no-unused-variable": true,
+        "no-var-keyword": true,
+        "no-var-requires": false,
+        "object-literal-sort-keys": [false, "ignore-case"],
+        "one-line": [
+            true,
+            "check-open-brace",
+            "check-catch",
+            "check-else",
+            "check-whitespace"
+        ],
+        "ordered-imports": false,
+        "quotemark": [
+            true,
+            "single",
+            "jsx-double"
+        ],
+        "radix": false,
+        "semicolon": [true, "never"],
+        "switch-default": true,
+        "trailing-comma": [
+            false,
+            {
+                "multiline": {
+                    "objects": "always",
+                    "arrays": "always",
+                    "functions": "never",
+                    "typeLiterals": "ignore"
+                },
+                "singleline": "never"
+            }
+        ],
+        "triple-equals": [
+            true,
+            "allow-null-check"
+        ],
+        "typedef-whitespace": [
+            true,
+            {
+                "call-signature": "nospace",
+                "index-signature": "nospace",
+                "parameter": "nospace",
+                "property-declaration": "nospace",
+                "variable-declaration": "nospace"
+            }
+        ],
+        "variable-name": false,
+        "whitespace": [
+            true,
+            "check-branch",
+            "check-decl",
+            "check-operator",
+            "check-separator",
+            "check-type",
+            "check-module",
+            "check-rest-spread",
+            "check-type-operator"
+        ]
+    }
+}
